Added modulus operation in PubVal and PrivVal Constructors
In a SNARK circuit, we need to use field arithmetic. When using larger numbers, they can "overflow".
This commit fixes a bug, e.g. when multiplying 2 large numbers, as it is automatically the case in the __pow__ routine:
Minimal Example which was previously rejected by the verifier:
`print(PubVal(2)**PubVal(3))`

This commit fixes some occurrences of such overflows, but there might still be similar issues in other places.